Fix Copilot review & move board scene creation to separate file
